Update angular-intro.js
There are cases where you need to evaluate a method to check if they have previously seen the intro. For example I set a cookie and any first time user login would not have that cookie set and the intro should auto start! I know this is not exactly like the options from intro.js but I am open to open a feature like this in another fashion if this solution is to intrusive.